Extract ClusterCache from Rule
